refactor: extract Discord gateway connection logic into dedicated module
Move WebSocket connection, heartbeat, identify/resume, and reconnect
logic from DiscordBridge (1252→1009 lines) into a new DiscordGateway
class (329 lines). Bridge delegates to gateway via dispatch callbacks.
